+++ src/Aws/CloudWatchLogs/CloudWatchLogsClient.php
@@ -14,22 +14,27 @@ use Guzzle\Service\Resource\Model;
  *
  * @method Model createLogGroup(array $args = array()) {@command CloudWatchLogs CreateLogGroup}
  * @method Model createLogStream(array $args = array()) {@command CloudWatchLogs CreateLogStream}
+ * @method Model deleteDestination(array $args = array()) {@command CloudWatchLogs DeleteDestination}
  * @method Model deleteLogGroup(array $args = array()) {@command CloudWatchLogs DeleteLogGroup}
  * @method Model deleteLogStream(array $args = array()) {@command CloudWatchLogs DeleteLogStream}
  * @method Model deleteMetricFilter(array $args = array()) {@command CloudWatchLogs DeleteMetricFilter}
  * @method Model deleteRetentionPolicy(array $args = array()) {@command CloudWatchLogs DeleteRetentionPolicy}
  * @method Model deleteSubscriptionFilter(array $args = array()) {@command CloudWatchLogs DeleteSubscriptionFilter}
+ * @method Model describeDestinations(array $args = array()) {@command CloudWatchLogs DescribeDestinations}
  * @method Model describeLogGroups(array $args = array()) {@command CloudWatchLogs DescribeLogGroups}
  * @method Model describeLogStreams(array $args = array()) {@command CloudWatchLogs DescribeLogStreams}
  * @method Model describeMetricFilters(array $args = array()) {@command CloudWatchLogs DescribeMetricFilters}
  * @method Model describeSubscriptionFilters(array $args = array()) {@command CloudWatchLogs DescribeSubscriptionFilters}
  * @method Model filterLogEvents(array $args = array()) {@command CloudWatchLogs FilterLogEvents}
  * @method Model getLogEvents(array $args = array()) {@command CloudWatchLogs GetLogEvents}
+ * @method Model putDestination(array $args = array()) {@command CloudWatchLogs PutDestination}
+ * @method Model putDestinationPolicy(array $args = array()) {@command CloudWatchLogs PutDestinationPolicy}
  * @method Model putLogEvents(array $args = array()) {@command CloudWatchLogs PutLogEvents}
  * @method Model putMetricFilter(array $args = array()) {@command CloudWatchLogs PutMetricFilter}
  * @method Model putRetentionPolicy(array $args = array()) {@command CloudWatchLogs PutRetentionPolicy}
  * @method Model putSubscriptionFilter(array $args = array()) {@command CloudWatchLogs PutSubscriptionFilter}
  * @method Model testMetricFilter(array $args = array()) {@command CloudWatchLogs TestMetricFilter}
+ * @method ResourceIteratorInterface getDescribeDestinationsIterator(array $args = array()) The input array uses the parameters of the DescribeDestinations operation
  * @method ResourceIteratorInterface getDescribeLogGroupsIterator(array $args = array()) The input array uses the parameters of the DescribeLogGroups operation
  * @method ResourceIteratorInterface getDescribeLogStreamsIterator(array $args = array()) The input array uses the parameters of the DescribeLogStreams operation
  * @method ResourceIteratorInterface getDescribeMetricFiltersIterator(array $args = array()) The input array uses the parameters of the DescribeMetricFilters operation