+++ src/Aws/CloudWatchLogs/Resources/cloudwatchlogs-2014-03-28.php
@@ -124,6 +124,54 @@ return array (
                 ),
             ),
         ),
+        'DeleteDestination' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'Logs_20140328.DeleteDestination',
+                ),
+                'destinationName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterException',
+                ),
+                array(
+                    'reason' => 'Returned if the specified resource does not exist.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+                array(
+                    'reason' => 'Returned if multiple requests to update the same resource were in conflict.',
+                    'class' => 'OperationAbortedException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
         'DeleteLogGroup' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -378,6 +426,56 @@ return array (
                 array(
                     'reason' => 'Returned if multiple requests to update the same resource were in conflict.',
                     'class' => 'OperationAbortedException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
+        'DescribeDestinations' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'DescribeDestinationsResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'Logs_20140328.DescribeDestinations',
+                ),
+                'DestinationNamePrefix' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+                'nextToken' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                ),
+                'limit' => array(
+                    'type' => 'numeric',
+                    'location' => 'json',
+                    'minimum' => 1,
+                    'maximum' => 50,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterException',
                 ),
                 array(
                     'reason' => 'Returned if the service cannot complete the request.',
@@ -787,6 +885,112 @@ return array (
                 ),
             ),
         ),
+        'PutDestination' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'PutDestinationResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'Logs_20140328.PutDestination',
+                ),
+                'destinationName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+                'targetArn' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                ),
+                'roleArn' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterException',
+                ),
+                array(
+                    'reason' => 'Returned if multiple requests to update the same resource were in conflict.',
+                    'class' => 'OperationAbortedException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
+        'PutDestinationPolicy' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'Logs_20140328.PutDestinationPolicy',
+                ),
+                'destinationName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 512,
+                ),
+                'accessPolicy' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterException',
+                ),
+                array(
+                    'reason' => 'Returned if multiple requests to update the same resource were in conflict.',
+                    'class' => 'OperationAbortedException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
         'PutLogEvents' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -1069,7 +1273,6 @@ return array (
                     'minLength' => 1,
                 ),
                 'roleArn' => array(
-                    'required' => true,
                     'type' => 'string',
                     'location' => 'json',
                     'minLength' => 1,
@@ -1154,6 +1357,44 @@ return array (
         'EmptyOutput' => array(
             'type' => 'object',
             'additionalProperties' => true,
+        ),
+        'DescribeDestinationsResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'destinations' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'Destination',
+                        'type' => 'object',
+                        'properties' => array(
+                            'destinationName' => array(
+                                'type' => 'string',
+                            ),
+                            'targetArn' => array(
+                                'type' => 'string',
+                            ),
+                            'roleArn' => array(
+                                'type' => 'string',
+                            ),
+                            'accessPolicy' => array(
+                                'type' => 'string',
+                            ),
+                            'arn' => array(
+                                'type' => 'string',
+                            ),
+                            'creationTime' => array(
+                                'type' => 'numeric',
+                            ),
+                        ),
+                    ),
+                ),
+                'nextToken' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
         ),
         'DescribeLogGroupsResponse' => array(
             'type' => 'object',
@@ -1406,6 +1647,36 @@ return array (
                 ),
             ),
         ),
+        'PutDestinationResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'destination' => array(
+                    'type' => 'object',
+                    'location' => 'json',
+                    'properties' => array(
+                        'destinationName' => array(
+                            'type' => 'string',
+                        ),
+                        'targetArn' => array(
+                            'type' => 'string',
+                        ),
+                        'roleArn' => array(
+                            'type' => 'string',
+                        ),
+                        'accessPolicy' => array(
+                            'type' => 'string',
+                        ),
+                        'arn' => array(
+                            'type' => 'string',
+                        ),
+                        'creationTime' => array(
+                            'type' => 'numeric',
+                        ),
+                    ),
+                ),
+            ),
+        ),
         'PutLogEventsResponse' => array(
             'type' => 'object',
             'additionalProperties' => true,
@@ -1461,6 +1732,12 @@ return array (
         ),
     ),
     'iterators' => array(
+        'DescribeDestinations' => array(
+            'input_token' => 'nextToken',
+            'output_token' => 'nextToken',
+            'limit_key' => 'limit',
+            'result_key' => 'destinations',
+        ),
         'DescribeLogGroups' => array(
             'input_token' => 'nextToken',
             'output_token' => 'nextToken',